Splits student view into tabs.
