+++ app/assets/javascripts/students/components/AddCodeTab.js
@@ -2,7 +2,7 @@
  * Component of `AddCoursesPage` that allows students to add more courses by
  * entering a code supplied by an admin. The student will then be enrolled in
  * the courses included as part of the code that he/she is not already enrolled
- * in.  
+ * in.
  */
 
 import React from 'react'
