+++ app/assets/javascripts/students/components/CourseRequestTab.js
@@ -1,17 +1,7 @@
-/**
- * Component of `AddCoursesPage` that allows students to request for courses
- * through selecting from the catalong. The other tab on the page uses codes.
- * Students can view and select any subset of the published courses and hit
- * submit, which will render a request viewable to all admins.
- */
-
 import React from 'react'
-import _ from 'underscore'
-
 import { APIRoutes } from '../../shared/routes'
 import request from '../../shared/requests/request'
-
-import StudentCourseRequestCard from './StudentCourseRequestCard'
+import _ from 'underscore'
 import Form from '../../shared/components/forms/Form'
 import Input from '../../shared/components/forms/Input'
 import SaveButton from '../../shared/components/widgets/SaveButton'
@@ -30,22 +20,13 @@ class CourseRequestTab extends React.Component {
         },
       },
       courses: [],
-      activeCourseIds: new Set(),
     }
 
-    this._mounted = false
-
-    this.generateRequests = this.generateRequests.bind(this)
-    this.switchCourseSelectedState = this.switchCourseSelectedState.bind(this)
+    this.generateRequests = _.bind(this.generateRequests, this)
   }
 
   componentDidMount() {
     this.getCourses()
-    this._mounted = true
-  }
-
-  componentWillUnmount() {
-    this._mounted = false
   }
 
   handleChange(attr, e) {
@@ -58,63 +39,43 @@ class CourseRequestTab extends React.Component {
     const path = APIRoutes.getPublishedCourses()
 
     request.get(path, (response) => {
-      this._mounted && this.setState({ courses: response.courses })
+      //_.every(enrolledList, (course) => { return course.progress === 100; }) ? "inline" : "none"
+      this.setState({ courses: response.courses })
+      console.log(response.courses)
     }, (error) => {
       console.log(error)
     })
   }
 
-  /**
-   * Removes `course` if already in this.state.activeCourseIds
-   * Otherwise adds `course` to this.state.activeCourseIds
-   */
-  switchCourseSelectedState(courseId) {
-    var activeCourseIds = this.state.activeCourseIds
-    if (activeCourseIds.has(courseId)) {
-      activeCourseIds.delete(courseId)
-    } else {
-      activeCourseIds.add(courseId)
-    }
-    this.setState({ activeCourseIds: activeCourseIds })
-  }
-
-  resetAllSelectedStates() {
-    this.setState({ activeCourseIds: new Set() })
-  }
-
   generateRequests(event, onSuccess, onFailure) {
     event.preventDefault()
     const path = APIRoutes.courseRequestPath()
+    var inputs = getInputToParams(this.state.formFields)
     var params = {
-      request: {
-        course_ids: [...this.state.activeCourseIds],
+      request_args: {
+        course_ids: JSON.stringify(this.state.formFields.course_ids.value)
       },
     }
     request.post(path, params, (response) => {
       onSuccess && onSuccess()
-      this.resetAllSelectedStates()
     }, (error) => {
       console.log(error)
       onFailure && onFailure()
     })
   }
 
+  renderFields() {
+    return (
+      _.pairs(this.state.formFields).map((values) => {
+        return <Input key={values[0]} {...values[1]} />
+      })
+    )
+  }
+
   renderCourses() {
-    if (_.isEmpty(this.state.courses)) {
-      return (
-        <div className='student-course-request-placeholder'>
-          <i className='fa fa-spinner fa-pulse save-button-icon'></i>
-          Loading all courses...
-        </div>
-      )
-    }
     return this.state.courses.map((value) => {
       return (
-        <StudentCourseRequestCard
-          course={value}
-          selected={this.state.activeCourseIds.has(value.id)}
-          updateSelected={this.switchCourseSelectedState}
-        />
+        <li key={value.id}>{value.name+" "+value.id}</li>
       )
     })
   }
@@ -123,15 +84,22 @@ class CourseRequestTab extends React.Component {
     return (
       <div>
         {this.renderCourses()}
-        <SaveButton
-          className='marginTop-md marginBot-xl submit-request-button'
-          text='Submit request'
-          onPress={this.generateRequests}
-          disabled={this.state.activeCourseIds.size === 0}
-        />
+        <Form
+          className='submit_request_ids_form'
+          id='submit_request_ids_form'
+          method='post'
+          action={this.props.action}
+        >
+          {this.renderFields()}
+          <SaveButton
+            text='Submit Requests'
+            onPress={this.generateRequests}
+          />
+        </Form>
       </div>
     )
   }
+
 }
 
 export default CourseRequestTab
